minor: 优化cc转移主机到默认模块代码
